♻️ [REFACTOR] : 페이지 별로 컴포넌트 폴더 재구성
