+++ src/components/Common/Badge/EmojiBadge/PaperListEmojiBadge.jsx
