+++ src/pages/PaperListPage/PaperListPage.js
@@ -1,31 +1,21 @@
 /* eslint-disable no-console */
 import React, { useState } from 'react';
 import * as S from './PaperListPage.style';
-import useRequest from './useRequest';
 import { Link } from 'react-router-dom';
-import PaperCard from '../../components/PaperCard';
-import ArrowButton from '../../components/ArrowButton';
-import PaperListSkeleton from '../../components/Skeleton/PaperListSkeleton';
+import PaperCard from './components/PaperCard';
+import ArrowButton from './components/ArrowButton/ArrowButton';
+import PaperListSkeleton from './components/Skeleton/PaperListSkeleton';
 import Header from '../../components/Common/Header/Header';
 import { PurpleButton } from '../../components/Common/PurpleButton/PurpleButton';
+import getRecipientData from '../../API';
 
 const PaperListPage = () => {
-  const { data: recentPaper, isLoading: isLoadingRecent } = useRequest({
-    options: {
-      url: 'recipients/',
-      method: 'get',
-    },
-  });
-
-  const { data: popularPaper, isLoading: isLoadingPopular } = useRequest({
-    options: {
-      url: 'recipients/',
-      method: 'get',
-      params: {
-        sort: 'like',
-      },
-    },
-  });
+  const {
+    getPopularPaperData,
+    isLoadingPopular,
+    getRecentPaperData,
+    isLoadingRecent,
+  } = getRecipientData();
 
   return (
     <>
@@ -33,12 +23,12 @@ const PaperListPage = () => {
       <S.Container>
         <PaperSection
           title="인기 롤링 페이퍼 🔥"
-          papers={popularPaper}
+          papers={getPopularPaperData}
           isLoading={isLoadingPopular}
         />
         <PaperSection
           title="최근에 만든 롤링 페이퍼 ⭐️"
-          papers={recentPaper}
+          papers={getRecentPaperData}
           isLoading={isLoadingRecent}
         />
       </S.Container>
