+++ src/pages/PaperListPage/components/ArrowButton/ArrowButton.style.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { COLORS } from '../../style/colorPalette';
+import { COLORS } from '../../../../style/colorPalette';
 
 export const ButtonContainer = styled.button`
   display: flex;
