+++ src/pages/PaperListPage/components/PaperCard/PaperCard.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import * as S from './PaperCard.style';
-import From from '../PaperListFromBadge';
-import Emoji from '../Common/PaperListEmojiBadge';
-import PatternPurple from '../../assets/icon/pattern_purple.svg';
-import PatternBeige from '../../assets/icon/pattern_beige.svg';
-import PatternBlue from '../../assets/icon/pattern_blue.svg';
-import PatternGreen from '../../assets/icon/pattern_green.svg';
+import From from '../../../../components/Common/Badge/FromBadge';
+import Emoji from '../../../../components/Common/Badge/EmojiBadge';
+import PatternPurple from '../../../../assets/icon/pattern_purple.svg';
+import PatternBeige from '../../../../assets/icon/pattern_beige.svg';
+import PatternBlue from '../../../../assets/icon/pattern_blue.svg';
+import PatternGreen from '../../../../assets/icon/pattern_green.svg';
 
 const BACKGROUND = {
   purple: { color: 'PURPLE_200', pattern: PatternPurple },
