+++ src/pages/PaperListPage/components/PaperCard/index.js
