+++ src/pages/PaperListPage/components/Skeleton/DelayComponent.js
