--- src/pages/PaperListPage/mock.js
@@ -1,454 +1,0 @@
-const mock = {
-  count: 10,
-  next: null,
-  previous: null,
-  results: [
-    {
-      id: 826,
-      name: '박지윤',
-      backgroundColor: 'green',
-      backgroundImageURL: null,
-      createdAt: '2023-12-04T01:45:16.017197Z',
-      messageCount: 5,
-      recentMessages: [
-        {
-          id: 2637,
-          recipientId: 826,
-          sender: 'test user10',
-          profileImageURL:
-            'https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8',
-          relationship: '지인',
-          content: '테스트',
-          font: '나눔명조',
-          createdAt: '2023-12-04T01:45:34.074951Z',
-        },
-        {
-          id: 2637,
-          recipientId: 826,
-          sender: 'test user10',
-          profileImageURL:
-            'https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8',
-          relationship: '지인',
-          content: '테스트',
-          font: '나눔명조',
-          createdAt: '2023-12-04T01:45:34.074951Z',
-        },
-        {
-          id: 2637,
-          recipientId: 826,
-          sender: 'test user10',
-          profileImageURL:
-            'https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8',
-          relationship: '지인',
-          content: '테스트',
-          font: '나눔명조',
-          createdAt: '2023-12-04T01:45:34.074951Z',
-        },
-        {
-          id: 2637,
-          recipientId: 826,
-          sender: 'test user10',
-          profileImageURL:
-            'https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8',
-          relationship: '지인',
-          content: '테스트',
-          font: '나눔명조',
-          createdAt: '2023-12-04T01:45:34.074951Z',
-        },
-        {
-          id: 2637,
-          recipientId: 826,
-          sender: 'test user10',
-          profileImageURL:
-            'https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8',
-          relationship: '지인',
-          content: '테스트',
-          font: '나눔명조',
-          createdAt: '2023-12-04T01:45:34.074951Z',
-        },
-      ],
-      reactionCount: 0,
-      topReactions: [
-        {
-          id: 28,
-          emoji: '🎀',
-          count: 14,
-        },
-        {
-          id: 31,
-          emoji: '💞',
-          count: 11,
-        },
-        {
-          id: 27,
-          emoji: '⭐️',
-          count: 14,
-        },
-      ],
-    },
-    {
-      id: 826,
-      name: '김김김',
-      backgroundColor: 'purple',
-      backgroundImageURL: null,
-      createdAt: '2023-12-04T01:45:16.017197Z',
-      messageCount: 1,
-      recentMessages: [
-        {
-          id: 2637,
-          recipientId: 826,
-          sender: 'test user10',
-          profileImageURL:
-            'https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8',
-          relationship: '지인',
-          content: '테스트',
-          font: '나눔명조',
-          createdAt: '2023-12-04T01:45:34.074951Z',
-        },
-      ],
-      reactionCount: 0,
-      topReactions: [
-        {
-          id: 27,
-          emoji: '😀',
-          count: 14,
-        },
-        {
-          id: 31,
-          emoji: '🥹',
-          count: 11,
-        },
-        {
-          id: 27,
-          emoji: '😀',
-          count: 14,
-        },
-      ],
-    },
-    {
-      id: 826,
-      name: '나나나',
-      backgroundColor: 'green',
-      backgroundImageURL: null,
-      createdAt: '2023-12-04T01:45:16.017197Z',
-      messageCount: 1,
-      recentMessages: [
-        {
-          id: 2637,
-          recipientId: 826,
-          sender: 'test user10',
-          profileImageURL:
-            'https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8',
-          relationship: '지인',
-          content: '테스트',
-          font: '나눔명조',
-          createdAt: '2023-12-04T01:45:34.074951Z',
-        },
-      ],
-      reactionCount: 0,
-      topReactions: [
-        {
-          id: 27,
-          emoji: '🎀',
-          count: 3,
-        },
-        {
-          id: 31,
-          emoji: '💜',
-          count: 12,
-        },
-        {
-          id: 27,
-          emoji: '🧡',
-          count: 22,
-        },
-      ],
-    },
-    {
-      id: 826,
-      name: '다다다',
-      backgroundColor: 'blue',
-      backgroundImageURL: null,
-      createdAt: '2023-12-04T01:45:16.017197Z',
-      messageCount: 2,
-      recentMessages: [
-        {
-          id: 2637,
-          recipientId: 826,
-          sender: 'test user10',
-          profileImageURL:
-            'https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8',
-          relationship: '지인',
-          content: '테스트',
-          font: '나눔명조',
-          createdAt: '2023-12-04T01:45:34.074951Z',
-        },
-        {
-          id: 2637,
-          recipientId: 826,
-          sender: 'test user10',
-          profileImageURL:
-            'https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8',
-          relationship: '지인',
-          content: '테스트',
-          font: '나눔명조',
-          createdAt: '2023-12-04T01:45:34.074951Z',
-        },
-      ],
-      reactionCount: 0,
-      topReactions: [
-        {
-          id: 27,
-          emoji: '🥹',
-          count: 143,
-        },
-        {
-          id: 31,
-          emoji: '💞',
-          count: 1,
-        },
-        {
-          id: 27,
-          emoji: '❓',
-          count: 14,
-        },
-      ],
-    },
-    {
-      id: 826,
-      name: '이름이 뭘까',
-      backgroundColor: 'beige',
-      backgroundImageURL: null,
-      createdAt: '2023-12-04T01:45:16.017197Z',
-      messageCount: 1,
-      recentMessages: [
-        {
-          id: 2637,
-          recipientId: 826,
-          sender: 'test user10',
-          profileImageURL:
-            'https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8',
-          relationship: '지인',
-          content: '테스트',
-          font: '나눔명조',
-          createdAt: '2023-12-04T01:45:34.074951Z',
-        },
-      ],
-      reactionCount: 0,
-      topReactions: [
-        {
-          id: 27,
-          emoji: '📝',
-          count: 1,
-        },
-        {
-          id: 31,
-          emoji: '📍',
-          count: 1,
-        },
-        {
-          id: 27,
-          emoji: '❓',
-          count: 14,
-        },
-      ],
-    },
-    {
-      id: 826,
-      name: '이르으으음',
-      backgroundColor: 'blue',
-      backgroundImageURL: null,
-      createdAt: '2023-12-04T01:45:16.017197Z',
-      messageCount: 1,
-      recentMessages: [
-        {
-          id: 2637,
-          recipientId: 826,
-          sender: 'test user10',
-          profileImageURL:
-            'https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8',
-          relationship: '지인',
-          content: '테스트',
-          font: '나눔명조',
-          createdAt: '2023-12-04T01:45:34.074951Z',
-        },
-      ],
-      reactionCount: 0,
-      topReactions: [
-        {
-          id: 27,
-          emoji: '🥹',
-          count: 3,
-        },
-        {
-          id: 31,
-          emoji: '⭐️',
-          count: 10,
-        },
-        {
-          id: 27,
-          emoji: '❓',
-          count: 14,
-        },
-      ],
-    },
-    {
-      id: 826,
-      name: '제니',
-      backgroundColor: 'beige',
-      backgroundImageURL: null,
-      createdAt: '2023-12-04T01:45:16.017197Z',
-      messageCount: 1,
-      recentMessages: [
-        {
-          id: 2637,
-          recipientId: 826,
-          sender: 'test user10',
-          profileImageURL:
-            'https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8',
-          relationship: '지인',
-          content: '테스트',
-          font: '나눔명조',
-          createdAt: '2023-12-04T01:45:34.074951Z',
-        },
-      ],
-      reactionCount: 0,
-      topReactions: [
-        {
-          id: 27,
-          emoji: '🩵',
-          count: 1,
-        },
-        {
-          id: 31,
-          emoji: '📸',
-          count: 10,
-        },
-        {
-          id: 27,
-          emoji: '❓',
-          count: 4,
-        },
-      ],
-    },
-    {
-      id: 826,
-      name: '로제',
-      backgroundColor: 'purple',
-      backgroundImageURL: null,
-      createdAt: '2023-12-04T01:45:16.017197Z',
-      messageCount: 1,
-      recentMessages: [
-        {
-          id: 2637,
-          recipientId: 826,
-          sender: 'test user10',
-          profileImageURL:
-            'https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8',
-          relationship: '지인',
-          content: '테스트',
-          font: '나눔명조',
-          createdAt: '2023-12-04T01:45:34.074951Z',
-        },
-      ],
-      reactionCount: 0,
-      topReactions: [
-        {
-          id: 27,
-          emoji: '🥹',
-          count: 143,
-        },
-        {
-          id: 31,
-          emoji: '💞',
-          count: 1,
-        },
-        {
-          id: 27,
-          emoji: '❓',
-          count: 14,
-        },
-      ],
-    },
-    {
-      id: 826,
-      name: '리사',
-      backgroundColor: 'green',
-      backgroundImageURL: null,
-      createdAt: '2023-12-04T01:45:16.017197Z',
-      messageCount: 1,
-      recentMessages: [
-        {
-          id: 2637,
-          recipientId: 826,
-          sender: 'test user10',
-          profileImageURL:
-            'https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8',
-          relationship: '지인',
-          content: '테스트',
-          font: '나눔명조',
-          createdAt: '2023-12-04T01:45:34.074951Z',
-        },
-      ],
-      reactionCount: 0,
-      topReactions: [
-        {
-          id: 27,
-          emoji: '🥹',
-          count: 143,
-        },
-        {
-          id: 31,
-          emoji: '💞',
-          count: 1,
-        },
-        {
-          id: 27,
-          emoji: '❓',
-          count: 14,
-        },
-      ],
-    },
-    {
-      id: 826,
-      name: '지수',
-      backgroundColor: 'purple',
-      backgroundImageURL: null,
-      createdAt: '2023-12-04T01:45:16.017197Z',
-      messageCount: 1,
-      recentMessages: [
-        {
-          id: 2637,
-          recipientId: 826,
-          sender: 'test user10',
-          profileImageURL:
-            'https://fastly.picsum.photos/id/311/200/200.jpg?hmac=CHiYGYQ3Xpesshw5eYWH7U0Kyl9zMTZLQuRDU4OtyH8',
-          relationship: '지인',
-          content: '테스트',
-          font: '나눔명조',
-          createdAt: '2023-12-04T01:45:34.074951Z',
-        },
-      ],
-      reactionCount: 0,
-      topReactions: [
-        {
-          id: 27,
-          emoji: '🥹',
-          count: 143,
-        },
-        {
-          id: 31,
-          emoji: '💞',
-          count: 1,
-        },
-        {
-          id: 27,
-          emoji: '❓',
-          count: 14,
-        },
-      ],
-    },
-  ],
-};
-
-export default mock;
